Simplify alignment logic
In case where at eof, do not treat zip data descriptor specially, other
than to add it to the calculation of space required. Thus if the data
descriptor would lead to crossing the boundary, we now always write a
padding frame, then a start-of-part, then the last data frame and data
descriptor in the next part.